% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_clocks.R
\docType{data}
\name{common_clocks}
\alias{common_clocks}
\title{Name of Common Clocks}
\format{
An object of class \code{list} of length 6.
}
\usage{
common_clocks
}
\description{
Common clocks names stored as a character vector
}
\keyword{datasets}
