% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_n_paths.R
\name{files_n_paths}
\alias{files_n_paths}
\title{Getting Files and Path in a Dir}
\usage{
files_n_paths(dir, ...)
}
\arguments{
\item{dir}{Directory}

\item{...}{Passed to \code{\link[=list.files]{list.files()}}}
}
\description{
Getting Files and Path in a Dir
}
