% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{impute.knn1}
\alias{impute.knn1}
\title{Wrapper for \code{\link[impute:impute.knn]{impute::impute.knn()}}}
\usage{
impute.knn1(X, feature_id, sample_id, k, maxp_prop, seed, na_loc, ...)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{feature_id}{A vector of feature IDs to subset rows of \code{X}.}

\item{sample_id}{A vector of sample IDs to subset columns of \code{X}.}

\item{k}{Integer specifying the number of nearest neighbors to use.}

\item{maxp_prop}{A numeric value specifying the proportion of features used per block.}

\item{seed}{An integer seed for random number generation.}

\item{na_loc}{A vector of location for amputation. Generated by \code{\link[=inject_na]{inject_na()}}.}

\item{...}{arguments passed to \code{\link[impute:impute.knn]{impute::impute.knn()}}}
}
\description{
Wrapper for \code{\link[impute:impute.knn]{impute::impute.knn()}}
}
