% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{impute_knn}
\alias{impute_knn}
\title{Wrapper for \code{\link[impute:impute.knn]{impute::impute.knn()}}}
\usage{
impute_knn(X, hash = NULL, k, maxp_prop, seed, na_loc, ...)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{hash}{hash of the saved object used to check hash.}

\item{k}{Integer specifying the number of nearest neighbors to use.}

\item{maxp_prop}{A numeric value specifying the proportion of features used per block.}

\item{seed}{An integer seed for random number generation.}

\item{na_loc}{A vector of location for amputation. Generated by \code{\link[=inject_na]{inject_na()}}.}

\item{...}{arguments passed to \code{\link[impute:impute.knn]{impute::impute.knn()}}}
}
\description{
Wrapper for \code{\link[impute:impute.knn]{impute::impute.knn()}}
}
