% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{imputePCA1}
\alias{imputePCA1}
\title{Wrapper for \code{\link[missMDA:methyl_imputePCA]{missMDA::methyl_imputePCA()}}}
\usage{
imputePCA1(X, feature_id, sample_id, ncp, seed, na_loc, ...)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{feature_id}{A vector of feature IDs to subset rows of \code{X}.}

\item{sample_id}{A vector of sample IDs to subset columns of \code{X}.}

\item{ncp}{number of PCs.}

\item{seed}{An integer seed for random number generation.}

\item{na_loc}{A vector of location for amputation. Generated by \code{\link[=inject_na]{inject_na()}}}

\item{...}{arguments passed to \code{\link[missMDA:methyl_imputePCA]{missMDA::methyl_imputePCA()}}}
}
\description{
Wrapper for \code{\link[missMDA:methyl_imputePCA]{missMDA::methyl_imputePCA()}}
}
