% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{inject_na}
\alias{inject_na}
\title{Inject Missing Values}
\usage{
inject_na(
  X,
  hash = NULL,
  prop,
  min_na,
  max_na,
  c_thresh = 0.9,
  r_thresh = 0.9,
  max_iter = 1000,
  transpose,
  debug = FALSE
)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{hash}{hash of the saved object used to check hash.}

\item{prop}{A numeric value between 0 and 1. Proportion of values to inject as missing in the subset of the matrix. Default is \code{0.05}.}

\item{min_na}{An integer. Minimum number of \code{NA} injected. Default is \code{50}. Used to calculate prediction performance.}

\item{max_na}{An integer. Maximum number of \code{NA} injected. Default is \code{5000}. Used to calculate prediction performance.}

\item{c_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per column. Default is \code{0.9}.}

\item{r_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per row. Default is \code{0.9}.}

\item{max_iter}{A positive integer. Maximum number of iterations for injecting missing values. Default is 1000.}

\item{transpose}{Transpose before NA injection or not. Default to FALSE}

\item{debug}{return extra object for debugging.}
}
\value{
A vector of indices indicating the locations of injected \code{NA} in the subset matrix.
}
\description{
Injects missing values \code{NA} into a subset of a matrix while adhering to column-wise and row-wise
missingness thresholds.
}
