% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is_unique}
\alias{is_unique}
\alias{is_unique.default}
\title{Check if a data.frame or vector contains unique values only}
\usage{
is_unique(obj)

\method{is_unique}{default}(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
boolean of yes unique or no not unique
}
\description{
Check if a data.frame or vector contains unique values only
}
