% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsave1.R
\name{qsave1}
\alias{qsave1}
\title{Modify qs::qsave By Saving to Temp First}
\usage{
qsave1(obj, file, overwrite = TRUE, ...)
}
\arguments{
\item{obj}{obj to serialize with qs}

\item{file}{destination}

\item{overwrite}{overwrite or not}

\item{...}{args to qsave}
}
\description{
Saving files to the hdd can create errors. Saving to temp first might create
less errors
}
