% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{save_sub_matrix}
\alias{save_sub_matrix}
\title{Save Submatrices to Files or Memory}
\usage{
save_sub_matrix(df, X, temp_path, progress)
}
\arguments{
\item{df}{data.frame}

\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{temp_path}{Storage location for matrices. Options:
\itemize{
\item \code{NULL}: Save in temporary directory (cleared at session end)
\item Character string: Save in specified directory path
\item \code{NA}: Store in memory (may significantly increase memory usage)
}}

\item{progress}{Logical. If \code{TRUE}, displays a progress bar for the tuning process. Default is \code{FALSE}.}
}
\description{
Takes a data frame containing feature and sample IDs, extracts corresponding
submatrices from a source matrix, and either saves them to disk or keeps them
in memory.
}
