% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{tune_impute_methyLImp2}
\alias{tune_impute_methyLImp2}
\title{Tune \code{\link[methyLImp2:methyLImp2]{methyLImp2::methyLImp2()}}}
\usage{
tune_impute_methyLImp2(
  X,
  group_sample,
  group_feature,
  hp = NULL,
  rep = 1,
  prop = 0.05,
  min_na,
  max_na,
  c_thresh = 0.9,
  r_thresh = 0.9,
  max_iter = 1000,
  parallel = FALSE,
  progress = FALSE,
  temp_path = NULL,
  ...
)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{group_sample}{A data frame with two columns: \code{sample_id} and \code{group}. Maps samples to their respective groups. Default is \code{NULL}.}

\item{group_feature}{A data frame with two columns: \code{feature_id} and \code{group}. Maps features to their respective groups. Default is \code{NULL}.}

\item{hp}{A data frame. Optional hyperparameter grid to be used for tuning. Default is \code{NULL}.}

\item{rep}{A vector of unique elements. The length of the vector is the times the tuning is repeated.}

\item{prop}{A numeric value between 0 and 1. Proportion of values to inject as missing in the subset of the matrix. Default is \code{0.05}.}

\item{min_na}{An integer. Minimum number of \code{NA} injected. Default is \code{50}. Used to calculate prediction performance.}

\item{max_na}{An integer. Maximum number of \code{NA} injected. Default is \code{5000}. Used to calculate prediction performance.}

\item{c_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per column. Default is \code{0.9}.}

\item{r_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per row. Default is \code{0.9}.}

\item{max_iter}{A positive integer. Maximum number of iterations for injecting missing values. Default is 1000.}

\item{parallel}{Logical. If \code{TRUE}, enables parallel processing using the \code{furrr} package. Default is \code{FALSE}.}

\item{progress}{Logical. If \code{TRUE}, displays a progress bar for the tuning process. Default is \code{FALSE}.}

\item{temp_path}{Storage location for matrices. Options:
\itemize{
\item \code{NULL}: Save in temporary directory (cleared at session end)
\item Character string: Save in specified directory path
\item \code{NA}: Store in memory (may significantly increase memory usage)
}}

\item{...}{Arguments passed to \code{\link[methyLImp2:methyLImp2]{methyLImp2::methyLImp2()}}}
}
\value{
A data frame containing the tuning parameters and a column called \code{tune}
that contains the tuning results for each parameter combination.\code{tune} can be used
for performance measurement (i.e. with \code{\link[yardstick:rmse]{yardstick::rmse()}}).
}
\description{
Tunes the parameters of the \code{\link[methyLImp2:methyLImp2]{methyLImp2::methyLImp2()}} method for imputing missing values.
This function evaluates the performance of different parameter combinations through repeated
\code{NA} injection and imputation, across various feature and sample groupings.
}
