% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{tune_imputePCA}
\alias{tune_imputePCA}
\title{Tune \code{\link[missMDA:methyl_imputePCA]{missMDA::methyl_imputePCA()}}}
\usage{
tune_imputePCA(
  X,
  group_sample,
  group_feature,
  hp,
  rep = 1,
  prop = 0.05,
  c_thresh = 0.9,
  r_thresh = 0.9,
  max_iter = 1000,
  parallel = FALSE,
  progress = FALSE,
  ...
)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{group_sample}{A data frame with two columns: \code{sample_id} and \code{group}. Maps samples to their respective groups. Default is \code{NULL}.}

\item{group_feature}{A data frame with two columns: \code{feature_id} and \code{group}. Maps features to their respective groups. Default is \code{NULL}.}

\item{hp}{A data frame. Optional hyperparameter grid to be used for tuning. Default is \code{NULL}.}

\item{rep}{A positive integer. The number of repeats for NA injection. Default is \code{1}.}

\item{prop}{A numeric value between 0 and 1. Proportion of values to inject as missing in the subset of the matrix. Default is \code{0.05}.}

\item{c_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per column. Default is \code{0.9}.}

\item{r_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per row. Default is \code{0.9}.}

\item{max_iter}{A positive integer. Maximum number of iterations for injecting missing values. Default is 1000.}

\item{parallel}{Logical. If \code{TRUE}, enables parallel processing using the \code{furrr} package. Default is \code{FALSE}.}

\item{progress}{Logical. If \code{TRUE}, displays a progress bar for the tuning process. Default is \code{FALSE}.}

\item{...}{Arguments passed to \code{\link[missMDA:methyl_imputePCA]{missMDA::methyl_imputePCA()}}}
}
\value{
A data frame containing the tuning parameters and a column called \code{tune}
that contains the tuning results for each parameter combination.\code{tune} can be used
for performance measurement (i.e. with \code{\link[yardstick:rmse]{yardstick::rmse()}}).
}
\description{
Tunes the parameters of the \code{\link[missMDA:methyl_imputePCA]{missMDA::methyl_imputePCA()}} method for imputing missing values.
This function evaluates the performance of different parameter combinations through repeated
NA injection and imputation, across various feature and sample groupings.
}
\examples{
X <- matrix(rnorm(100), nrow = 10, ncol = 10)
row.names(X) <- 1:10
colnames(X) <- 1:10
group_sample <- data.frame(sample_id = colnames(X), group = rep(1:2, each = 5))
group_feature <- data.frame(feature_id = rownames(X), group = rep(1:2, each = 5))
hp <- data.frame(ncp = 1)
tune_imputePCA(
  X,
  group_sample,
  group_feature,
  hp,
  rep = 2,
  parallel = FALSE,
  progress = TRUE,
  c_thresh = 1,
  r_thresh = 1
)
}
