% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{inject_na}
\alias{inject_na}
\title{Inject Missing Values}
\usage{
inject_na(
  X,
  feature_id,
  sample_id,
  prop,
  c_thresh = 0.9,
  r_thresh = 0.9,
  max_iter = 1000
)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{feature_id}{A vector of feature IDs to subset rows of \code{X}.}

\item{sample_id}{A vector of sample IDs to subset columns of \code{X}.}

\item{prop}{A numeric value between 0 and 1. Proportion of values to inject as missing in the subset of the matrix. Default is \code{0.05}.}

\item{c_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per column. Default is \code{0.9}.}

\item{r_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per row. Default is \code{0.9}.}

\item{max_iter}{A positive integer. Maximum number of iterations for injecting missing values. Default is 1000.}
}
\value{
A vector of indices indicating the locations of injected \code{NA} in the subset matrix.
}
\description{
Injects missing values \code{NA} into a subset of a matrix while adhering to column-wise and row-wise
missingness thresholds.
}
\examples{
X <- matrix(1:1000, nrow = 100, ncol = 10)
colnames(X) <- 1:10
rownames(X) <- 1:100
feature_id <- rownames(X)[1:5]
sample_id <- colnames(X)[1:5]
inject_na(X, feature_id, sample_id, prop = 0.1, c_thresh = 1, r_thresh = 1)
}
