% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_imputation.R
\name{tune_prep}
\alias{tune_prep}
\title{Setup Grid Search for Tuning Parameters}
\usage{
tune_prep(
  X,
  group_sample = NULL,
  group_feature = NULL,
  rep = 1,
  hp = NULL,
  prop = 0.05,
  c_thresh = 0.9,
  r_thresh = 0.9,
  max_iter = 1000,
  transpose = FALSE
)
}
\arguments{
\item{X}{A numeric matrix. The input data to be imputed. Rows correspond to features, and columns correspond to samples.}

\item{group_sample}{A data frame with two columns: \code{sample_id} and \code{group}. Maps samples to their respective groups. Default is \code{NULL}.}

\item{group_feature}{A data frame with two columns: \code{feature_id} and \code{group}. Maps features to their respective groups. Default is \code{NULL}.}

\item{rep}{A positive integer. The number of repeats for NA injection. Default is \code{1}.}

\item{hp}{A data frame. Optional hyperparameter grid to be used for tuning. Default is \code{NULL}.}

\item{prop}{A numeric value between 0 and 1. Proportion of values to inject as missing in the subset of the matrix. Default is \code{0.05}.}

\item{c_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per column. Default is \code{0.9}.}

\item{r_thresh}{A numeric value between 0 and 1. Maximum allowed proportion of missing values per row. Default is \code{0.9}.}

\item{max_iter}{A positive integer. Maximum number of iterations for injecting missing values. Default is 1000.}

\item{transpose}{Transpose before NA injection or not. Default to FALSE}
}
\value{
A data frame
}
\description{
Biological assays often require methods to impute missing data, which may involve hyperparameters.
Additionally, multiple methods may be applied. This function sets up a grid search to tune
imputation methods by injecting \code{NA} multiple times, imputing the data,
and comparing the imputed data to the original data before the \code{NA} injection. The function
also sets up tuning stratified by feature groups (e.g., chromosomes) and sample groups
(e.g., sample batches).
}
\examples{
X <- matrix(rnorm(100), nrow = 10, ncol = 10)
row.names(X) <- 1:10
colnames(X) <- 1:10
group_sample <- data.frame(sample_id = colnames(X), group = rep(1:2, each = 5))
group_feature <- data.frame(feature_id = rownames(X), group = rep(1:2, each = 5))
tune_prep(X, group_sample, group_feature, rep = 3, prop = 0.2, c_thresh = 1, r_thresh = 1)
}
